Sat, Nov 19, 2022, 8:21 PM +00:00

+++ runoff/runoff.c
@@ -163,7 +163,7 @@ bool print_winner(void)
     int index2;
     for (int i = 0; i < candidate_count; i++)
     {
-        if ((float)candidates[i].votes > max)
+        if (candidates[i].votes > max)
         {
             max = candidates[i].votes;
             index2 = i;
